+++ inbas-reporting-tool/src/main/webapp/js/i18n/en.js
@@ -261,6 +261,8 @@ module.exports = {
         'report.attachments.create.description-label': 'Description',
         'report.attachments.create.description-tooltip': 'Optional description of the attachment',
         'report.attachments.table.reference': 'Reference',
+        'report.eccairs.button.label': 'ECCAIRS report',
+        'report.eccairs.button.tooltip': 'View latest ECCAIRS report for this report',
 
         'report.occurrence.category.label': 'Occurrence category',
         'occurrencereport.title': 'Occurrence report',
@@ -344,6 +346,7 @@ module.exports = {
         'statistics.type.general': 'General',
         'statistics.type.eventTypes': 'Event types',
         'statistics.type.audit': 'Audit',
+        'statistics.type.safetyIssue': 'Safety issue',
 
         'search.loading': 'Searching...',
         'search.title': 'Search results',